Implemented solution for CodeEval's Digit Statistics challenge https://www.codeeval.com/open_challenges/144/
This one was a tough one, I started with an O(n) algorithm, which obviously was way to slow for big exponents.
That made me think there had to be a better way, so I found faster pow algorithms which run in O(log(n))
But since we need all exponents, they would need O(n*log(n)), which is slower then my O(n) algorithm.

So I figured there had to be an O(1) algorithm, while showering I was thinking about this and started playing with numbers.
So I started with base 5 and kept multiplying it by 5 which made me realize that the last digit is always 5.
So I figured that other numbers might also have a repeating sequence for their end digits.

Later during the day, while I was at a calculator I could soon test that theory :-)
And I found repeating sequences for all numbers from 1 to 9 which means I could solve this problem in O(1)