Implemented solution for HackerRanks Gemstone challenge
Calculate the Set of elements which is used in all stones
For fun, I store the Alphabet of the Stone, inside of a 32 bit Integer
So that evaluation will be just an AND over all stones :)
Which for more fun I do inline, while reading in the stones.
So I never have to store more then one stone.

https://www.hackerrank.com/challenges/gem-stones